Create useFirebaseOnce hook to fetch values and refetch predictably
