+++ src/utils/hooksHelper.js
@@ -1,4 +1,4 @@
-import { useRef, useEffect } from "react"
+import { useCallback, useRef, useEffect, useState } from "react"
 
 function usePrevious(value) {
   const ref = useRef()
@@ -8,4 +8,58 @@ function usePrevious(value) {
   return ref.current
 }
 
-export { usePrevious }
+/**
+ *
+ * Fetches data from firebase .once('value)
+ * @param {*} dbRef String: ref to fetch from
+ * @param {*} shouldRefetchData Boolean: condition when data should be refetched
+ * @returns
+ * {
+ *   data: Array: [{ key, ...values }]
+ *   dataObj: Object {key: val, ...}
+ *   loading: Boolean
+ * }
+ */
+function useFirebaseOnceHook(dbRef, shouldRefetchData) {
+  const [data, setData] = useState(undefined)
+  const [dataObj, setDataObj] = useState(undefined)
+  const [loading, setLoading] = useState(false)
+
+  const fetchData = useCallback(() => {
+    setLoading(true)
+    dbRef.once("value").then((s) => {
+      // console.log("&& fetched", s.val())
+      const vals = s.val()
+      if (vals) {
+        setData(
+          Object.keys(vals).map((key) => ({ [key]: vals[key], ...vals[key] }))
+        )
+        setDataObj(s.val())
+      } else {
+        setData([])
+        setDataObj({})
+      }
+      setLoading(false)
+    })
+  }, [dbRef])
+
+  // init fetch
+  useEffect(() => {
+    if (data === undefined && !loading) {
+      // console.log("initial fetch data")
+      fetchData()
+    }
+  }, [loading, data, fetchData])
+
+  // refetch data when condition is passed
+  useEffect(() => {
+    if (shouldRefetchData) {
+      // console.log("refetch data")
+      fetchData()
+    }
+  }, [shouldRefetchData, fetchData])
+
+  return { data, dataObj, loading }
+}
+
+export { usePrevious, useFirebaseOnceHook }
